src: move top-level bitstreams into `src/top`
