+++ .github/workflows/main.yml
@@ -13,7 +13,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm run src/example_usb_audio/top.py
+          pdm usb_audio
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -30,7 +30,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm run src/example_dsp/top.py --dsp-core nco
+          pdm dsp --dsp-core nco
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -47,7 +47,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm run src/example_dsp/top.py --sim --dsp-core nco
+          pdm dsp --sim --dsp-core nco
         working-directory: gateware
 
   ubuntu-bitstream-dsp-midicv:
@@ -60,7 +60,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm run src/example_dsp/top.py --dsp-core midicv
+          pdm dsp --dsp-core midicv
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -84,7 +84,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm run src/xbeam/top.py
+          pdm xbeam
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -108,7 +108,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm run src/selftest/top.py
+          pdm selftest
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
@@ -133,7 +133,7 @@ jobs:
       - run: yosys --version
       - run: |
           pdm install
-          pdm run src/poly/top.py
+          pdm polysyn
         working-directory: gateware
       - uses: actions/upload-artifact@v4
         with:
