--- gateware/src/boot/fw/src/handlers.rs
@@ -1,1 +1,0 @@
-../../../rs/handlers.rs
