+++ gateware/src/tiliqua/eurorack_pmod.py
@@ -14,8 +14,6 @@ from amaranth.lib.wiring        import In, Out
 from amaranth.lib.fifo          import AsyncFIFO
 from amaranth.lib.cdc           import FFSynchronizer
 
-from example_usb_audio.util     import EdgeToPulse
-
 from amaranth_future            import fixed
 
 WIDTH = 16
@@ -84,6 +82,29 @@ class AudioStream(wiring.Component):
                     self.eurorack_pmod.sample_o.eq(dac_fifo.r_data),
                 ]
                 m.next = 'READ'
+
+        return m
+
+# TODO: move this to another util lib
+class EdgeToPulse(Elaboratable):
+    """
+    each rising edge of the signal edge_in will be
+    converted to a single clock pulse on pulse_out
+    """
+    def __init__(self):
+        self.edge_in          = Signal()
+        self.pulse_out        = Signal()
+
+    def elaborate(self, platform) -> Module:
+        m = Module()
+
+        edge_last = Signal()
+
+        m.d.sync += edge_last.eq(self.edge_in)
+        with m.If(self.edge_in & ~edge_last):
+            m.d.comb += self.pulse_out.eq(1)
+        with m.Else():
+            m.d.comb += self.pulse_out.eq(0)
 
         return m
