+++ gateware/src/tiliqua/scope.py
@@ -2,28 +2,14 @@
 #
 # SPDX-License-Identifier: CERN-OHL-S-2.0
 
-"""
-Advanced version of `example_vectorscope` that includes an SoC and menu system.
-Rasterizes X/Y (audio channel 0, 1) and color (audio channel 3) to a simulated CRT.
-"""
-
-import logging
-import os
-import sys
-
 from amaranth                                    import *
 from amaranth.lib                                import wiring, data, stream
 from amaranth.lib.wiring                         import In, Out, flipped, connect
 
 from amaranth_soc                                import csr
 
-from amaranth_future                             import fixed
-
-from tiliqua                                     import eurorack_pmod, dsp
-from tiliqua.tiliqua_soc                         import TiliquaSoc
-from tiliqua.cli                                 import top_level_cli
-
-from example_vectorscope.top                     import Stroke
+from tiliqua                                     import dsp
+from tiliqua.raster                              import Stroke
 
 class VectorTracePeripheral(wiring.Component):
 
@@ -238,71 +224,3 @@ class ScopeTracePeripheral(wiring.Component):
 
         return m
 
-class XbeamSoc(TiliquaSoc):
-    def __init__(self, *, firmware_path, dvi_timings, **kwargs):
-
-        # don't finalize the CSR bridge in TiliquaSoc, we're adding more peripherals.
-        super().__init__(firmware_path=firmware_path, dvi_timings=dvi_timings, audio_192=True,
-                         audio_out_peripheral=False, finalize_csr_bridge=False, **kwargs)
-
-        # scope stroke bridge from audio stream
-        fb_size = (self.video.fb_hsize, self.video.fb_vsize)
-
-        # WARN: TiliquaSoc ends at 0x00000900
-        self.vector_periph_base = 0x00001000
-        self.scope_periph_base  = 0x00001100
-
-        self.vector_periph = VectorTracePeripheral(
-            fb_base=self.video.fb_base,
-            fb_size=fb_size,
-            bus_dma=self.psram_periph,
-            video_rotate_90=self.video_rotate_90)
-        self.csr_decoder.add(self.vector_periph.bus, addr=self.vector_periph_base, name="vector_periph")
-
-        self.scope_periph = ScopeTracePeripheral(
-            fb_base=self.video.fb_base,
-            fb_size=fb_size,
-            bus_dma=self.psram_periph,
-            video_rotate_90=self.video_rotate_90)
-        self.csr_decoder.add(self.scope_periph.bus, addr=self.scope_periph_base, name="scope_periph")
-
-        # now we can freeze the memory map
-        self.finalize_csr_bridge()
-
-    def elaborate(self, platform):
-
-        m = Module()
-
-        m.submodules += self.vector_periph
-
-        m.submodules += self.scope_periph
-
-        m.submodules += super().elaborate(platform)
-
-        pmod0 = self.pmod0_periph.pmod
-
-        m.submodules.astream = astream = eurorack_pmod.AudioStream(pmod0)
-
-        self.scope_periph.source = astream.istream
-
-        with m.If(self.scope_periph.soc_en):
-            wiring.connect(m, astream.istream, self.scope_periph.i)
-        with m.Else():
-            wiring.connect(m, astream.istream, self.vector_periph.i)
-
-        m.d.comb += [
-            astream.ostream.valid.eq(astream.istream.valid & astream.istream.ready),
-            astream.ostream.payload.eq(astream.istream.payload),
-        ]
-
-        # Memory controller hangs if we start making requests to it straight away.
-        with m.If(self.permit_bus_traffic):
-            m.d.sync += self.vector_periph.en.eq(1)
-            m.d.sync += self.scope_periph.en.eq(1)
-
-        return m
-
-
-if __name__ == "__main__":
-    this_path = os.path.dirname(os.path.realpath(__file__))
-    top_level_cli(XbeamSoc, path=this_path)
