+++ gateware/src/top/bootloader/.gitignore
