+++ gateware/src/top/bootloader/fw/Cargo.toml
@@ -4,9 +4,9 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-tiliqua-hal = { path="../../rs/hal", default-features = false }
-tiliqua-lib = { path = "../../rs/lib" }
-tiliqua-pac = { path="../../rs/pac", default-features = false, features = ["critical-section", "vexriscv"] }
+tiliqua-hal = { path="../../../rs/hal", default-features = false }
+tiliqua-lib = { path = "../../../rs/lib" }
+tiliqua-pac = { path="../../../rs/pac", default-features = false, features = ["critical-section", "vexriscv"] }
 
 log = "0.4.17"
 embedded-hal = "=1.0.0"
@@ -19,7 +19,6 @@ fastrand = { version = "2.1.0", default-features = false }
 heapless = "0.8.0"
 strum_macros = "0.26.4"
 strum = {version = "0.25.0", features = ["derive"], default-features=false}
-colorsys = { version = "*", default-features = false }
 
 [profile.release]
 codegen-units = 1
