+++ gateware/src/top/dsp/sim_dsp_core.cpp
