+++ gateware/src/top/polysyn/fw/Cargo.toml
@@ -4,9 +4,9 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-tiliqua-hal = { path="../../rs/hal", default-features = false }
-tiliqua-lib = { path = "../../rs/lib" }
-tiliqua-pac = { path="../../rs/pac", default-features = false, features = ["critical-section", "vexriscv"] }
+tiliqua-hal = { path="../../../rs/hal", default-features = false }
+tiliqua-lib = { path = "../../../rs/lib" }
+tiliqua-pac = { path="../../../rs/pac", default-features = false, features = ["critical-section", "vexriscv"] }
 
 log = "0.4.17"
 embedded-hal = "=1.0.0"
