+++ gateware/src/top/selftest/fw/src/handlers.rs
@@ -1,0 +1,1 @@
+../../../../rs/handlers.rs
