+++ gateware/src/top/selftest/fw/src/lib.rs
