+++ gateware/src/top/selftest/fw/src/main.rs
