+++ gateware/src/top/selftest/fw/src/opts.rs
