+++ gateware/src/top/selftest/sim.cpp
