+++ gateware/src/top/usb_audio/channels_to_usb_stream.py
@@ -6,7 +6,7 @@ from amaranth             import *
 from amaranth.build       import Platform
 from amaranth.lib.fifo    import SyncFIFO
 from luna.gateware.stream import StreamInterface
-from example_usb_audio.util import connect_fifo_to_stream
+from util                 import connect_fifo_to_stream
 
 class ChannelsToUSBStream(Elaboratable):
     def __init__(self, max_nr_channels=2, sample_width=24, max_packet_size=256):
