+++ gateware/src/top/xbeam/fw/src/opts.rs
